requirements.txt
attrs==22.1.0
autopep8==1.7.0
bleach==5.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
commonmark==0.9.1
docutils==0.19
five9==0.0.3b192
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.0
grpcio-status==1.49.0
idna==3.4
importlib-metadata==4.12.0
isodate==0.6.1
jaraco.classes==3.2.2
keyring==23.9.1
lxml==4.9.1
more-itertools==8.14.0
numpy==1.23.3
pandas==1.5.0
pkginfo==1.8.3
platformdirs==2.5.2
properties==0.6.1
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.2.1
readme-renderer==37.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
rsa==4.9
six==1.16.0
toml==0.10.2
twine==4.0.1
typing_extensions==4.3.0
urllib3==1.26.12
webencodings==0.5.1
websocket-client==1.4.1
zeep==4.1.0
zipp==3.8.1
